requirements.txt
Django
python-dotenv==0.10.1
PyJWT
django_tables2
django-filter
